feat(single-wager) refactoring
