+++ src/components/LoginForm.js
@@ -1,0 +1,52 @@
+import React from 'react'
+import { Link } from "react-router-dom";
+import { Grid, Header, Form, Message, Segment, Button } from 'semantic-ui-react'
+
+const LoginForm = (props) => (
+  <div className='login-form'>
+    <style>{`
+      body > div,
+      body > div > div,
+      body > div > div > div.login-form {
+        height: 100%;
+      }
+    `}</style>
+    <Grid
+      textAlign='center'
+      style={{ height: '100%' }}
+
+    >
+      <Grid.Column style={{ maxWidth: 450 }}>
+        <Header as='h2' color='black' textAlign='center'>
+          Log in Below
+        </Header>
+        <Form onSubmit={props.login} size='large'>
+          <Segment stacked>
+            <Form.Input
+              fluid
+              icon='user'
+              iconPosition='left'
+              placeholder='E-mail address'
+              name="email"
+            />
+            <Form.Input
+              fluid
+              icon='lock'
+              iconPosition='left'
+              placeholder='Password'
+              type='password'
+              name="password"
+            />
+
+            <Button type='submit' color='black' fluid size='large'>Login</Button>
+          </Segment>
+        </Form>
+        <Message>
+          New to us? <Link to='/signup'>Sign Up</Link>
+        </Message>
+      </Grid.Column>
+    </Grid>
+  </div>
+)
+
+export default LoginForm
